converted accural values to support floats

+++ internal/structs/withdraw.go
@@ -1,7 +1,7 @@
 package structs
 
 type Withdraw struct {
-	Order       string `json:"order"`
-	Sum         int    `json:"sum"`
-	ProcessedAt string `json:"processed_at,omitempty"`
+	Order       string  `json:"order"`
+	Sum         float32 `json:"sum"`
+	ProcessedAt string  `json:"processed_at,omitempty"`
 }
